requirements.txt
alembic==1.4.1
Babel==2.8.0
blinker==1.4
Bootstrap-Flask==1.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cx-Freeze==6.1
dominate==2.5.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.0
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.7
visitor==0.1.3
Werkzeug==1.0.0
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==2.2.1
